modules/location_generator_service/requirements.txt
kafka-python==2.0.2
grpcio==1.47.2
grpcio-tools==1.47.2